Tree is a virtual AI Assistant coded entirely on python. This application uses the likes of modules like pyttsx3, sapi5 and voiceID.
pyttsx3 : A python library that will help us to convert text to speech. In short, it is a text-to-speech library.
It works offline, and it is compatible with Python 2 as well as Python 3.

sapi5
Microsoft developed speech API.
Helps in synthesis and recognition of voice.

VoiceId
Voice id helps us to select different voices.
voice[0].id = Male voice voice[1].id = Female voice